3-4から6norm
