+++ ForSubmit/Makefile
@@ -6,7 +6,7 @@
 #    By: yokitaga <yokitaga@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/10/08 11:19:11 by yokitaga          #+#    #+#              #
-#    Updated: 2023/11/25 15:46:44 by yokitaga         ###   ########.fr        #
+#    Updated: 2023/11/25 16:08:32 by yokitaga         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -25,6 +25,7 @@ SRCS 			= $(addprefix $(SRCS_DIR),\
 			3_3_parse_read_data.c \
 			3_4_parse_read_data.c \
 			3_5_parse_read_data.c \
+			3_6_parse_read_data.c \
 			4_1_init_other_data.c \
 			4_2_init_other_data.c \
 			5_raycaster_draw.c \
