+++ ForSubmit/srcs/3_4_parse_read_data.c
@@ -6,7 +6,7 @@
 /*   By: yokitaga <yokitaga@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 12:31:24 by yokitaga          #+#    #+#             */
-/*   Updated: 2023/11/25 16:06:17 by yokitaga         ###   ########.fr       */
+/*   Updated: 2023/11/25 16:07:59 by yokitaga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,61 +80,4 @@ bool	check_leftside_wall(char	**map)
 		j++;
 	left_end_wall_index = j;
 	return (check_leftside_wall_for_norm(map, left_end_wall_index));
-}
-
-bool	check_rightside_wall(char	**map)
-{
-	size_t	i;
-	size_t	j;
-	size_t	right_end_wall_index;
-	size_t	next_len;
-
-	i = 0;
-	j = 0;
-	while (map[0][j] != '\0')
-		j++;
-	j--;
-	while (map[0][j] == 'X')
-		j--;
-	right_end_wall_index = j;
-	while (map[i + 1] != NULL)
-	{
-		next_len = ft_strlen(map[i + 1]);
-		if (right_end_wall_index + 1 > next_len)
-		{
-			while (right_end_wall_index > 0)
-			{
-				if (map[i][right_end_wall_index - 1] == '1')
-					right_end_wall_index--;
-				else
-					break ;
-			}
-			if (right_end_wall_index + 1 > next_len)
-				return (false);
-		}
-		if (map[i + 1][right_end_wall_index] == 'X')
-		{
-			while (right_end_wall_index > 0 && \
-				map[i][right_end_wall_index - 1] == '1' \
-				&& map[i + 1][right_end_wall_index] == 'X')
-				right_end_wall_index--;
-			if (map[i + 1][right_end_wall_index] != '1')
-				return (false);
-			else
-				i++;
-		}
-		else if (map[i + 1][right_end_wall_index] == '0')
-		{
-			while (map[i][right_end_wall_index + 1] == '1' && \
-				map[i + 1][right_end_wall_index] == '0')
-				right_end_wall_index++;
-			if (map[i + 1][right_end_wall_index] != '1')
-				return (false);
-			else
-				i++;
-		}
-		else
-			i++;
-	}
-	return (true);
 }
